Transition to custom elements.
There are many benefits to using custom elements. First, it allows to
separate concerns between parsers and editing elemnets.

It also avoid the use of toString() on HTMLElement which could create
issues with other framework

Also, I have this evil plan to have a sort of diffing algorithm to be
able to reuse HTML elements which would remove some flickers like when
rebuilding the document with multiple images.