Revert to not using Custom elements + parsers refactor
Custom Elements would be awesome to use but they aren't ready for prime
time. The current polyfills address most of the spec but not everything
and I needed to have a specific feature that only worked on Chrome
natively. While I could have come up with a solution, I though that
maybe this was a sign to revert. Even though this makes me work more.

I also took the time to revisit the parsers and look at a solution that
would be much easier to work with. I think this current version is much
closer to what I would expect as parsers.

Documentations on how to create your own parser will be coming soon when
I finalize the API for them.